# docs/api.openapi.yaml
...
openapi: 3.0.3
info:
  title: RPG 게임 로그 적재 & 집계 API
  version: 2.0.0
  description: |
    제약이 있는 HTTP 전송 환경(120회/분, 요청 4MB, timeout 30초)에서
    게임 이벤트를 유실 없이 적재하고, 운영 지표를 조회하는 API.
    - 멱등성: 같은 event_id 재전송 시 중복 저장되지 않음 (duplicate로 집계)
    - 시각: 모든 시각은 UTC ISO8601(ms). 지표는 occurred_at 기준.
      모든 기간은 반개구간 [start 00:00, end+1일 00:00)
    - 모든 지표는 조회 시점 적재 데이터 기준의 가변 결과 (늦은 이벤트로 변경 가능)
servers:
  - url: http://localhost:3000
    description: 로컬 개발 환경 (실배포에서는 HTTPS 443, TLS는 프록시 종료)
...
security:
  - bearerAuth: []
paths:
  /api/v1/event-batches:
    post:
      tags: [ingestion]
      operationId: ingestEventBatch
      summary: 이벤트 배치 적재
      description: |
        한 요청 = 한 배치(1~500건, 최대 4MB). 배치 구조가 유효하면 모든 이벤트가
        검증 실패하더라도 200 + per-event 결과로 응답한다(부분 실패 정책).
        응답은 DB 커밋 이후에만 반환된다 (성공 응답 = 저장 완료).
        카운트 불변식: received = accepted + rejected, accepted = stored + duplicate,
        order_duplicate ⊆ stored.
        전송측은 HTTP 상태가 아니라 응답 본문을 파싱해 stored/duplicate/order_duplicate로
        확인된 이벤트만 outbox에서 삭제해야 한다 (design.md §4.2).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventBatchRequest'
      responses:
        '200':
          description: 적재 완료 (부분 실패 포함 — rejected 배열 확인)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventBatchResponse'
        '400': { $ref: '#/components/responses/MalformedRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/InstanceMismatch' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/StorageUnavailable' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 적재는 인스턴스별 API 키, 조회는 관리자 키
...
  schemas:
    EventType:
      type: string
      enum:
        - session_login
        - session_logout
        - level_up
        - exp_gain
        - monster_kill
        - quest_complete
        - item_acquire
        - item_use
        - currency_change
        - shop_purchase
        - boss_clear
        - map_enter
        - death
    EventEnvelope:
      type: object
      required:
        - instance_id
        - event_id
        - event_type
        - user_id
        - character_id
        - session_id
        - channel_id
        - payload
        - occurred_at
      properties:
        instance_id: { type: string, format: uuid }
        event_id:
          type: string
          format: uuid
          description: 멱등성 키. 같은 값 재전송 시 중복 저장되지 않음
        event_type: { $ref: '#/components/schemas/EventType' }
        user_id: { type: integer, format: int64, example: 1001 }
        character_id: { type: integer, format: int64, example: 2001 }
        session_id: { type: string, minLength: 1, maxLength: 64 }
        channel_id: { type: string, minLength: 1, maxLength: 64 }
        payload:
          type: object
          description: |
            타입별 상세. shop_purchase만 필드 단위 엄격 검증(ShopPurchasePayload),
            그 외 타입은 object 여부만 확인 후 원본 보존 (design.md §8.1)
        occurred_at:
          type: string
          format: date-time
          description: |
            ISO8601 UTC, ms 정밀도. 시간대 지정자(Z 또는 ±hh:mm) 필수 —
            없으면 rejected(INVALID_ENVELOPE)
          example: '2026-07-27T04:15:22.123Z'
...
    EventBatchRequest:
      type: object
      required: [batch_id, sent_at, events]
      properties:
        batch_id:
          type: string
          format: uuid
          description: 요청 추적·응답 대조용 (멱등성 키 아님 — 멱등성은 event_id 단위)
        sent_at: { type: string, format: date-time }
        events:
          type: array
          minItems: 1
          maxItems: 500
          items: { $ref: '#/components/schemas/EventEnvelope' }
    EventBatchResponse:
      type: object
      required:
        - batch_id
        - received_count
        - accepted_count
        - stored_count
        - duplicate_count
        - order_duplicate_count
        - rejected_count
        - rejected
      description: |
        불변식: received = accepted + rejected / accepted = stored + duplicate /
        order_duplicate ⊆ stored
      properties:
        batch_id: { type: string, format: uuid }
        received_count: { type: integer, description: 배치에 담겨 온 이벤트 수 }
        accepted_count: { type: integer, description: 형식이 유효한 이벤트 수 (중복 포함) }
        stored_count: { type: integer, description: 이번 요청으로 원본이 새로 저장된 수 }
        duplicate_count:
          type: integer
          description: event_id PK 충돌로 건너뛴 수. 내용 동일성 검증 결과가 아님
        order_duplicate_count:
          type: integer
          description: 원본은 저장됐으나 order_id UNIQUE 충돌로 결제 파생만 생략된 수 (stored의 부분집합)
        rejected_count: { type: integer }
        rejected:
          type: array
          items:
            type: object
            required: [index, code, message]
            properties:
              index: { type: integer, description: 배치 내 위치 (0-base) }
              event_id: { type: string, format: uuid }
              code: { type: string, example: INVALID_PAYLOAD }
              message:
                type: string
                maxLength: 200
                description: 서버 정의 문구. 사용자 입력을 반사하지 않음
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: INVALID_DATE_RANGE }
            message: { type: string, example: start must be less than or equal to end }
            request_id: { type: string, example: req-7f8dd7 }
  responses:
    MalformedRequest:
      description: |
        JSON 파싱 실패 또는 배치 필수 필드 누락 (개별 이벤트 검증 실패는 200 + rejected).
        Content-Length 없는 청크형 비-JSON 전송의 안전 거부 포함 (design.md §2.3)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    Unauthorized:
      description: API 키 누락 또는 무효
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    InstanceMismatch:
      description: API 키에 매핑된 instance_id와 배치 내 이벤트의 instance_id 불일치
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    PayloadTooLarge:
      description: 요청 본문 4MB 초과 — 전송측은 배치를 분할 재전송 (design.md §4.2)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    RateLimited:
      description: |
        서버 측 요청 한도 초과 — 적재 경로에 키(=인스턴스) 단위 고정 1분 창
        (기본 120회/분, env RATE_LIMIT_PER_MINUTE)으로 적용된다 [A-25].
        인증(401/403)을 통과한 요청만 카운트하며, 429로 거절된 요청은
        어떤 이벤트도 저장하지 않는다. (design-aws.md §5)
      headers:
        Retry-After:
          description: 재시도까지 대기 시간(초) — 전송측은 jitter를 더해 대기
          schema: { type: integer, example: 5 }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    StorageUnavailable:
      description: |
        DB 장애·내부 데드라인 초과 등 일시적 오류 — 재시도 가능(503).
        적재는 트랜잭션 롤백 후 반환되며 전송측은 재시도한다.
        지표 조회의 스냅샷 트랜잭션 만료(P2028)·DB 연결 오류(P1001/P1002)도
        이 응답으로 매핑된다.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }